optimize docker layers
Before 2 big (>2GiB) layers was created and was dependent of the
presentation source code (ie. needed to be rebuild if the presentation
code changed).

Now we have the dependencies in one big layer, but the presentation in
another. So if the presentation change but its dependencies remain the
same, only the presentation layer will be rebuild and uploaded.